requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-on-heroku==1.1.2
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-on-heroku==1.1.2
django-storages==1.11.1
djangorestframework==3.12.4
dropbox==11.13.1
gunicorn==20.1.0
idna==2.10
ply==3.11
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.2.0
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.4
django-import-export==2.5.0
django-material-admin==1.8.5
django-on-heroku==1.1.2
django-storages==1.11.1
djangorestframework==3.12.4
dropbox==11.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
ply==3.11
psycopg2-binary==2.9.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
tablib==3.0.0
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
